Split sema.rs monolith into focused modules
The 6,469-line sema.rs has been split into 8 focused modules:

- mod.rs (889 lines): Core Sema struct and public API (analyze_all,
  analyze_all_bodies, gather_declarations)
- airgen.rs (4,119 lines): AIR generation from RIR (analyze_function,
  analyze_inst, and related helpers)
- items.rs (570 lines): Declaration gathering (register_type_names,
  resolve_declarations, struct/enum processing)
- context.rs (360 lines): Analysis context types (LocalVar, ParamInfo,
  VariableMoveState, AnalysisContext)
- types_helper.rs (291 lines): Type resolution helpers (resolve_type,
  format_type_name, is_type_copy, array type management)
- builtins.rs (143 lines): Built-in type injection (inject_builtin_types,
  is_builtin_string, builtin helpers)
- strings.rs (23 lines): String literal deduplication
- tests.rs (186 lines): Unit tests

This improves code organization and maintainability by grouping related
functionality together. Each module has a focused responsibility.

Implements rue-rxws.
