+++ crates/rue-air/src/sema/tests.rs
@@ -1,0 +1,186 @@
+//! Tests for semantic analysis.
+
+use rue_error::{CompileErrors, ErrorKind, MultiErrorResult, PreviewFeatures};
+use rue_lexer::Lexer;
+use rue_parser::Parser;
+use rue_rir::AstGen;
+
+use crate::inst::{AirInstData, AirRef};
+use crate::types::Type;
+
+use super::{Sema, SemaOutput};
+
+fn compile_to_air(source: &str) -> MultiErrorResult<SemaOutput> {
+    let lexer = Lexer::new(source);
+    let (tokens, interner) = lexer.tokenize().map_err(CompileErrors::from_error)?;
+    let parser = Parser::new(tokens, interner);
+    let (ast, mut interner) = parser.parse()?;
+
+    let astgen = AstGen::new(&ast, &mut interner);
+    let rir = astgen.generate();
+
+    let sema = Sema::new(&rir, &mut interner, PreviewFeatures::new());
+    sema.analyze_all()
+}
+
+#[test]
+fn test_analyze_simple_function() {
+    let output = compile_to_air("fn main() -> i32 { 42 }").unwrap();
+    let functions = &output.functions;
+
+    assert_eq!(functions.len(), 1);
+    assert_eq!(functions[0].name, "main");
+
+    let air = &functions[0].air;
+    assert_eq!(air.return_type(), Type::I32);
+    assert_eq!(air.len(), 2); // Const + Ret
+}
+
+#[test]
+fn test_analyze_addition() {
+    let output = compile_to_air("fn main() -> i32 { 1 + 2 }").unwrap();
+
+    let air = &output.functions[0].air;
+    assert_eq!(air.return_type(), Type::I32);
+    // Const(1) + Const(2) + Add + Ret = 4 instructions
+    assert_eq!(air.len(), 4);
+
+    // Check that add instruction exists with correct type
+    let add_inst = air.get(AirRef::from_raw(2));
+    assert!(matches!(add_inst.data, AirInstData::Add(_, _)));
+    assert_eq!(add_inst.ty, Type::I32);
+}
+
+#[test]
+fn test_analyze_all_binary_ops() {
+    // Test that all binary operators compile correctly
+    assert!(compile_to_air("fn main() -> i32 { 1 + 2 }").is_ok());
+    assert!(compile_to_air("fn main() -> i32 { 1 - 2 }").is_ok());
+    assert!(compile_to_air("fn main() -> i32 { 1 * 2 }").is_ok());
+    assert!(compile_to_air("fn main() -> i32 { 1 / 2 }").is_ok());
+    assert!(compile_to_air("fn main() -> i32 { 1 % 2 }").is_ok());
+}
+
+#[test]
+fn test_analyze_negation() {
+    let output = compile_to_air("fn main() -> i32 { -42 }").unwrap();
+
+    let air = &output.functions[0].air;
+    // Const(42) + Neg + Ret = 3 instructions
+    assert_eq!(air.len(), 3);
+
+    let neg_inst = air.get(AirRef::from_raw(1));
+    assert!(matches!(neg_inst.data, AirInstData::Neg(_)));
+    assert_eq!(neg_inst.ty, Type::I32);
+}
+
+#[test]
+fn test_analyze_complex_expr() {
+    let output = compile_to_air("fn main() -> i32 { (1 + 2) * 3 }").unwrap();
+
+    let air = &output.functions[0].air;
+    // Const(1) + Const(2) + Add + Const(3) + Mul + Ret = 6 instructions
+    assert_eq!(air.len(), 6);
+
+    // Check that result is multiplication
+    let mul_inst = air.get(AirRef::from_raw(4));
+    assert!(matches!(mul_inst.data, AirInstData::Mul(_, _)));
+}
+
+#[test]
+fn test_analyze_let_binding() {
+    let output = compile_to_air("fn main() -> i32 { let x = 42; x }").unwrap();
+
+    assert_eq!(output.functions.len(), 1);
+    assert_eq!(output.functions[0].num_locals, 1);
+
+    let air = &output.functions[0].air;
+    // Const(42) + StorageLive + Alloc + Block([StorageLive], Alloc) + Load + Block([alloc block], Load) + Ret = 7 instructions
+    assert_eq!(air.len(), 7);
+
+    // Check storage_live instruction
+    let storage_live_inst = air.get(AirRef::from_raw(1));
+    assert!(matches!(
+        storage_live_inst.data,
+        AirInstData::StorageLive { slot: 0 }
+    ));
+
+    // Check alloc instruction
+    let alloc_inst = air.get(AirRef::from_raw(2));
+    assert!(matches!(
+        alloc_inst.data,
+        AirInstData::Alloc { slot: 0, .. }
+    ));
+
+    // Check load instruction
+    let load_inst = air.get(AirRef::from_raw(4));
+    assert!(matches!(load_inst.data, AirInstData::Load { slot: 0 }));
+
+    // Check block instruction groups the alloc with the load
+    let block_inst = air.get(AirRef::from_raw(5));
+    assert!(matches!(block_inst.data, AirInstData::Block { .. }));
+}
+
+#[test]
+fn test_analyze_let_mut_assignment() {
+    let output = compile_to_air("fn main() -> i32 { let mut x = 10; x = 20; x }").unwrap();
+
+    let air = &output.functions[0].air;
+    // Const(10) + StorageLive + Alloc + Block([StorageLive], Alloc) + Const(20) + Store + Load + Block([alloc block, Store], Load) + Ret = 9 instructions
+    assert_eq!(air.len(), 9);
+
+    // Check store instruction
+    let store_inst = air.get(AirRef::from_raw(5));
+    assert!(matches!(
+        store_inst.data,
+        AirInstData::Store { slot: 0, .. }
+    ));
+
+    // Check block instruction groups statements
+    let block_inst = air.get(AirRef::from_raw(7));
+    assert!(matches!(block_inst.data, AirInstData::Block { .. }));
+}
+
+#[test]
+fn test_undefined_variable() {
+    let result = compile_to_air("fn main() -> i32 { x }");
+    assert!(result.is_err());
+    let errors = result.unwrap_err();
+    assert_eq!(errors.len(), 1);
+    assert!(matches!(
+        errors.iter().next().unwrap().kind,
+        ErrorKind::UndefinedVariable(_)
+    ));
+}
+
+#[test]
+fn test_assign_to_immutable() {
+    let result = compile_to_air("fn main() -> i32 { let x = 10; x = 20; x }");
+    assert!(result.is_err());
+    let errors = result.unwrap_err();
+    assert_eq!(errors.len(), 1);
+    assert!(matches!(
+        errors.iter().next().unwrap().kind,
+        ErrorKind::AssignToImmutable(_)
+    ));
+}
+
+#[test]
+fn test_multiple_variables() {
+    let output = compile_to_air("fn main() -> i32 { let x = 10; let y = 20; x + y }").unwrap();
+
+    assert_eq!(output.functions[0].num_locals, 2);
+}
+
+#[test]
+fn test_empty_block_evaluates_to_unit() {
+    // Empty block should evaluate to () and not panic
+    let output = compile_to_air("fn main() { let _x: () = {}; }").unwrap();
+
+    let air = &output.functions[0].air;
+    // Should have a UnitConst instruction for the empty block
+    let has_unit_const = air
+        .iter()
+        .any(|(_, inst)| matches!(inst.data, AirInstData::UnitConst));
+    assert!(has_unit_const, "Empty block should produce UnitConst");
+}
